# src/swagger/swagger.json
{
	"basePath": "/api/tnm/v1",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"ChargingScheduleSection": {
			"properties": {
				"time": {
					"type": "number",
					"format": "double"
				},
				"limit": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"time",
				"limit"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ChargingSchedule": {
			"properties": {
				"recurring": {
					"type": "string",
					"enum": [
						"weekly",
						"daily",
						"none"
					]
				},
				"sections": {
					"items": {
						"$ref": "#/definitions/ChargingScheduleSection"
					},
					"type": "array"
				}
			},
			"required": [
				"recurring",
				"sections"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "ncpa-api",
		"version": "1.0.0",
		"description": "Reference API for NCPA",
		"license": {
			"name": "UNLICENSED"
		}
	},
	"paths": {
...
		"/schedules/chargepoint": {
			"get": {
				"operationId": "GetSchedule",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ChargingSchedule"
						}
					}
				},
				"description": "jwt scopes: `operator` `",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "SetSchedule",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ChargingSchedule"
						}
					}
				},
				"description": "jwt scopes: `operator` `",
				"tags": [
					"operator"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "chargingSchedule",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ChargingSchedule"
						}
					},
					{
						"in": "query",
						"name": "clientId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		},
		"jwtAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "authorization"
		}
	}
}